tools: Move GenericVisitor to visitors.py
