+++ devito/tools/__init__.py
@@ -2,3 +2,4 @@ from devito.tools.utils import *  # noqa
 from devito.tools.memoization import *  # noqa
 from devito.tools.mixins import *  # noqa
 from devito.tools.validators import *  # noqa
+from devito.tools.visitors import *  # noqa
